JSON and UrlEncoded request middleware now encode body conditionally
These middleware now pay attention to "Content-Type" of the request.
Each will only encode the body if the content-type value matches.

When a body is present but no Content-Type, the middleware higher on
the stack will use its encoding.

This makes it possible to keep both middlewares on the stack and choose
encoding type by Content-Type switching.