Automatic parallel manager detection.
This enables the use of `#in_parallel` without manually passing in a parallel manager instance.

If the adapter you're using has a parallel manager, it will work as expected. If it doesn't, it'll work as normal sequential requests.

This is a bit of a dirty trick to figure out which middleware being loaded is an adapter by check if they respond to `#setup_parallel_manager`.