rewrite test script in bash
The HTTP server is now started in a subprocess, not a thread. Running it
in a Ruby thread had all sorts of problems on different Rubies, such as
random failures on Rubinius.

The new script also avoids starting the HTTP server if not needed, e.g.
if filtering arguments have been used to select only the non-adapter tests.

The script dumps the server log to stdout if any tests failed on Travis.

Avoids auto-starting the debugger because it blocks for input in Rubinius
https://travis-ci.org/lostisland/faraday/jobs/5956815