prevent Typhoeus from using "Expect: 100-continue" on PUT
It seems whether the request method for Typhoeus request is `:put` or
"PUT" decides whether it will use "Expect: 100-continue" when performing
that request. In our test suite, such PUTs would fail randomly against
WEBrick which doesn't have stable support for 100 Continue. This seems
to be a problem with WEBrick, not Typhoeus.

This also speeds up these requests, which WEBrick was slow to process
even if they've succeeded.

"Expect: 100-continue" is a mode where the client sends this header, but
doesn't send the body of the request until the server inspects the
headers and decides it wants to allow the request by sending
"HTTP/1.1 100 Continue" back.