Adding support for more SSL options for net/http.
I considered gating with a respond_to? to support older versions of
Ruby where net/http does not support all of the attributes, but doing so
would make Faraday silently ignore those values instead of notifying the
developer that they are trying to do something with an older version of
Ruby that it does not support.