Update ENV Keys docs to use `options` rather than `request`
The documentation for the ENV key isn’t `request`. Rather it is `options`. Since this isn’t very well documented for proxy settings it was unclear how to implement proxy settings. Changing to `options` makes it clear as it can be mapped back to the documentation for `timeout` as well.