Add test coverage for POST with Retry middleware
There was a fix put in bc82276 (#296) that preserved the body of a
POST request that was being retried in the Retry middleware.
Unfortunately, there was no test added that ensures that this
functionality continues to work.

This commit adds the test to ensure that it continues to work in the
future.