[httpclient] Fixes default certificate store not having default paths.
This came up while debugging an issue with the latest Berkshelf
release. It seems that the net_http.rb adapter did this in the exact
same way (calls #set_default_paths on the X509::Store) but for some
reason it was neglected when httpclient.rb adapter was written.

It is expectation that, by default, the system certificate store is utilized.