Fix EM::HTTP SSL patch to trust certificate verified by trusted intermediates
The Ubuntu team [recently released an update to the `ca-certificates`
package](https://launchpadlibrarian.net/242531582/ca-certificates_20141019ubuntu0.12.04.1_20160104ubuntu0.12.04.1.diff.gz)
that removes several root certificates, including one that is used to
sign the root of the chain for `login.mailchimp.com` (GTE CyberTrust Global Root).

One of the intermediaries on the certificate chain (Baltimore CyberTrust Root)
is included in Ubuntu's/Mozilla's ca-certificates bundle, and should be enough
to verify that the certificate is valid.

The patched version of `em-http` that is distributed with
faraday requires all certificates in the chain to be trusted before it'll verify
the connection. This causes ssl verification for login.mailchimp.com to fail.