Parse headers from aggregated proxy requests/responses
Prior to this commit when connecting to a proxy with auth, faraday
adapters would cause faraday to fail by first trying to connect without
auth and then connection again with auth, faraday could not handle the
response headers from the aggregate response of the two requests.
This commit addresses the above issue by changing faradays parse
function for headers to handle the case where there are multiple
repsonse headers by attempting to parse the last set of response
headers.