Support ssl-verify and ca-file option for Patron
The patron adapter didn't support the ssl: {verify: false} option nor
the ca_file. This was also the reason why the tests running on the
local server failed.

The adapter now supports both options. As Patron does not allow setting
the session.insecure option to true and supply a ca_file at the same
time, we need to only set either the insecure or the session.cacert
option.