Restore backward-compatibility layer for errors
Faraday has been maintaining a backward-compatibility layer for errors
in the `Faraday::Error::` "namespace". These were removed with v0.16.0,
which caused issues for dependents that use semantic versioning
constraints.

It would be best if these were deprecated with a warning message and
then dropped in v1.0 in order to make sure transitive dependencies do
not break.

See octokit/octokit.rb#1154