Replace the Net::HTTP adapter by a separate gem
The Net::HTTP adapter is extracted into its own gem. The idea behind this change is to create an ecosystem of different packages that you can use with Faraday, so that Faraday doesn’t have to ship with the packages itself.

For now all the Net::HTTP tests are still in place to verify that everything is working correctly.