Add strict_mode to Test::Stubs
Sometimes, users might want to check whether stubbed requests exactly
match the parameters and headers. In this case, they expect
`Test::Stubs` to raise an error when they add a new parameter in a
request. This patch introduces `strict_mode` for the purpose of the
expectation. With `strict_mode`, `Test::Stubs` tries to check all a
path, parameters, and headers exactly.