Respect the `params_encoder` in `Faraday::Adapter::Test`
Faraday provides a feature to let its callers configure
`params_encoder`, but `Faraday::Adapter::Test` had always created test
stubs with `#parse_nested_query` (`NestedParamsEncoder`). So, we had not
able to test such a connection with `Faraday::Adapter::Test`. This might
be inconvenient for users who want to use non-default `params_encoder`,
such as `Faraday::FlatParamsEncoder`.

This patch changed the adapter to use the same `params_encoder` as its
connection in order to solve the problem.