Make `Headers#[]=` set a value as String
`Faraday::Utils::Headers#[]=` has converted a passed value into a String
separated with `", "` if the value is an array. On the other hand, `#[]=`
had stored the value as is if the value is not an array.

This is the previous behavior:

```ruby
h = Faraday::Utils::Headers.new
h[:length] = 123
h[:length] # => 123

h[:length] = [123, 234]
h[:length] # => "123, 234"
```

I think this is a bit confusing, and it should consistently store a
value as string.

With this patch, we will get the headers with all stored values as
string.

```ruby
h = Faraday::Utils::Headers.new
h[:length] = 123
h[:length] # => "123"

h[:length] = [123, 234]
h[:length] # => "123, 234"
```

It might be a breaking change, but I think it's worth discussing.